Finalizado Cliente y DireccionFEL excepto controller

+++ src/main/java/com/techxperts/erp/cliente/model/Cliente.java
@@ -61,7 +61,7 @@ public class Cliente {
     private boolean bloquearCredito;
 
     @Column(name = "limite_credito")
-    private Double limiteCredito;
+    private Integer limiteCredito;
 
     @Column(name = "dias_credito")
     private Integer diasCredito;
